+++ docs/references/changelog.md
@@ -11,6 +11,12 @@ versioning][semver].
 
 ## Upcoming
 
+### Added
+
+- `kele-get` buffers now have a dedicated major mode, `kele-get-mode`.
+- `kele-get` buffers now print a header detailing, for the resource under
+  display: the **context** it was fetched from; and the **time of retrieval**
+
 ### Changed
 
 - Made `yaml-mode` a non-optional dependency